Optimize function resolving by moving it to the target process
No matter how fast we make the exports enumeration we will still have to
transfer about 12 MB of data for a typical iOS app. Thus it is a better
idea to move the processing to the target app, as it'll take a lot less
time even if there might be limited processing power.