Refactor to enable automatic signal bindings
This gives us the Script#destroyed signal which we need for solving
the RPC deadlock. It also gets rid of technical debt where a ton of
duplicated error-prone code was required per signal binding. There was
even a bug where every single one was unreffing the wrong Python object.
Oops.